DOMA-2617 moved render number to Renders.tsx
